+++ Hamsters/Hamsters/Extensions/String+.swift
@@ -12,3 +12,12 @@ extension String {
         return self.trimmingCharacters(in: .whitespacesAndNewlines)
     }
 }
+
+extension String {
+    var convertStringToPage: [DailyRecordPage] {
+        self.split(separator: " ")
+            .compactMap { Int(String($0)) }
+            .sorted{ $0 < $1 }
+            .compactMap { DailyRecordPage(rawValue: $0) }
+    }
+}
