+++ Hamsters/Hamsters/Scenes/Tab/RecordTab/DailyRecord/CaffeineCheckView.swift
@@ -93,6 +93,7 @@ struct CaffeineCheckView: View {
             NextButton(title: "다음", isActive: .constant(true)) {
                 dailyRecordViewModel.goToNextPage()
             }
+            .padding(.bottom, 40)
 //            DailyRecordNextButton(pageNumber: $pageNumber, isActiveRecord: .constant(true), title: "다음")
         }
     }
