+++ Hamsters/Hamsters/Scenes/Tab/RecordTab/DailyRecord/DailyCompleteView.swift
@@ -44,12 +44,13 @@ struct DailyCompleteView: View {
             NextButton(title: "다음", isActive: .constant(true)) {
                 dailyRecordViewModel.goToNextPage()
             }
+            .padding(.bottom, 40)
 //            DailyRecordNextButton(pageNumber: $pageNumber, isActiveRecord: $isActiveRecord, title: "완료")
             
         }
     }
 }
 
-//#Preview {
-//    DailyCompleteView(pageNumber: .constant(11))
-//}
+#Preview {
+    DailyCompleteView(dailyRecordViewModel: DailyRecordViewModel())
+}
