+++ Hamsters/Hamsters/Scenes/Tab/RecordTab/DailyRecord/DrinkCheckView.swift
@@ -74,6 +74,7 @@ struct DrinkCheckView: View {
             NextButton(title: "다음", isActive: .constant(true)) {
                 dailyRecordViewModel.goToNextPage()
             }
+            .padding(.bottom, 40)
             
 //            DailyRecordNextButton(pageNumber: $pageNumber, isActiveRecord: .constant(true), title: "다음")
         }
