+++ Hamsters/Hamsters/Scenes/Tab/RecordTab/DailyRecord/MenstruationCheckView.swift
@@ -54,6 +54,7 @@ struct MenstruationCheckView: View {
             NextButton(title: "다음", isActive: .constant(true)) {
                 dailyRecordViewModel.goToNextPage()
             }
+            .padding(.bottom, 40)
 //            DailyRecordNextButton(pageNumber: $pageNumber, isActiveRecord:.constant(true), title: "다음")
             
         }
