+++ Hamsters/Hamsters/Scenes/Tab/RecordTab/DailyRecord/SleepingTimeView.swift
@@ -112,6 +112,7 @@ struct SleepingTimeView: View {
                 }
                 .padding(.bottom, 30)
             }
+            
         }
     }
