+++ Hamsters/Hamsters/Scenes/Tab/RecordTab/DailyRecord/SmokingCheckView.swift
@@ -73,6 +73,7 @@ struct SmokingCheckView: View {
             NextButton(title: "다음", isActive: .constant(true)) {
                 dailyRecordViewModel.goToNextPage()
             }
+            .padding(.bottom, 40)
             
 //            DailyRecordNextButton(pageNumber: $pageNumber, isActiveRecord: .constant(true), title: "다음")
         }
